analytics/requirements.txt
# Analytics Backend Dependencies
# Python 3.11+

Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
gunicorn==21.2.0

